# internal/api/server/server.go
// Package server implements the daemon's local HTTP/JSON API. It is served over
// the IPC transport (Unix socket / named pipe), not a network port. All
// responses are JSON; errors use a consistent envelope.
package server

import (
	"encoding/json"
	"log/slog"
	"net/http"

	"github.com/shruggietech/go-schedule/internal/buildinfo"
	"github.com/shruggietech/go-schedule/internal/enrollment"
	"github.com/shruggietech/go-schedule/internal/events"
	"github.com/shruggietech/go-schedule/internal/logbus"
	"github.com/shruggietech/go-schedule/internal/store"
)

// Scheduler is the subset of the engine the API needs. It may be nil in tests
// that exercise only persistence-backed endpoints.
type Scheduler interface {
	Reload()
	RunNow(taskID string) error
}

// NotificationDispatcher accepts coalesced wake requests for durable work.
type NotificationDispatcher interface {
	Wake()
}

// Server holds dependencies and the route mux.
type Server struct {
	store        *store.Store
	sched        Scheduler
	broker       *events.Broker
	logs         *logbus.Ring
	logPath      string
	runtime      RuntimeInfoResponse
	log          *slog.Logger
	notify       NotificationDispatcher
	mcpHTTP      MCPHTTPManager
	mux          *http.ServeMux
	resolveActor func(*http.Request) (string, error)
	enrollment   *enrollment.Service
}

// RuntimeInfoResponse identifies the daemon's effective local storage paths.
// These values come from the configuration used by the running daemon, not
// from client-side defaults.
type RuntimeInfoResponse struct {
	DataDir      string `json:"data_dir"`
	DatabasePath string `json:"database_path"`
	ConfigPath   string `json:"config_path,omitempty"`
	LogPath      string `json:"log_path"`
	LockPath     string `json:"lock_path"`
}

// New constructs a Server and registers routes. sched, broker, and logs may be
// nil (e.g. in tests that exercise only persistence-backed endpoints).
func New(st *store.Store, sched Scheduler, broker *events.Broker, logs *logbus.Ring, logPath string, log *slog.Logger) *Server {
	return NewWithRuntimeInfo(st, sched, broker, logs, logPath, RuntimeInfoResponse{}, log)
}

// NewWithRuntimeInfo constructs a Server with authoritative daemon storage
// metadata for GET /v1/runtime-info.
func NewWithRuntimeInfo(st *store.Store, sched Scheduler, broker *events.Broker, logs *logbus.Ring, logPath string, runtime RuntimeInfoResponse, log *slog.Logger) *Server {
	s := &Server{store: st, sched: sched, broker: broker, logs: logs, logPath: logPath, runtime: runtime, log: log, mux: http.NewServeMux(), enrollment: enrollment.New(st)}
	s.resolveActor = func(*http.Request) (string, error) {
		actor, err := st.LocalActor()
		return actor.ID, err
	}
	s.routes()
	return s
}

// Handler returns the HTTP handler for the API.
func (s *Server) Handler() http.Handler { return http.HandlerFunc(s.authorizeAndAudit) }

// SetActorResolver supplies authenticated actor identity for a transport. The
// local transport uses the built-in operating-system actor by default.
func (s *Server) SetActorResolver(resolver func(*http.Request) (string, error)) {
	if resolver != nil {
		s.resolveActor = resolver
	}
}

// SetNotificationDispatcher connects channel tests and committed policy work
// to the outbound runtime without making handlers perform network requests.
func (s *Server) SetNotificationDispatcher(dispatcher NotificationDispatcher) { s.notify = dispatcher }

// SetMCPHTTPManager connects runtime-only localhost MCP lifecycle control.
func (s *Server) SetMCPHTTPManager(manager MCPHTTPManager) { s.mcpHTTP = manager }

func (s *Server) routes() {
	s.mux.HandleFunc("GET /v1/health", s.handleHealth)
	s.mux.HandleFunc("GET /v1/manifest", s.handleManifest)
	s.mux.HandleFunc("PATCH /v1/manifest", s.handleRenameManifest)
	s.mux.HandleFunc("POST /v1/manifest/reset", s.handleResetManifest)
	s.mux.HandleFunc("GET /v1/runtime-info", s.handleRuntimeInfo)
	s.mux.HandleFunc("GET /v1/mcp/http", s.handleMCPHTTPStatus)
	s.mux.HandleFunc("POST /v1/mcp/http/enable", s.handleMCPHTTPEnable)
	s.mux.HandleFunc("POST /v1/mcp/http/rotate", s.handleMCPHTTPRotate)
	s.mux.HandleFunc("POST /v1/mcp/http/disable", s.handleMCPHTTPDisable)

	s.mux.HandleFunc("GET /v1/tasks", s.handleListTasks)
	s.mux.HandleFunc("POST /v1/tasks", s.handleCreateTask)
	s.mux.HandleFunc("GET /v1/tasks/{id}", s.handleGetTask)
	s.mux.HandleFunc("PATCH /v1/tasks/{id}", s.handleUpdateTask)
	s.mux.HandleFunc("DELETE /v1/tasks/{id}", s.handleDeleteTask)
	s.mux.HandleFunc("POST /v1/tasks/{id}/enable", s.handleEnableTask)
	s.mux.HandleFunc("POST /v1/tasks/{id}/disable", s.handleDisableTask)
	s.mux.HandleFunc("POST /v1/tasks/{id}/run-now", s.handleRunNow)

	s.mux.HandleFunc("GET /v1/triggers", s.handleListTriggers)
	s.mux.HandleFunc("POST /v1/triggers", s.handleCreateTrigger)
	s.mux.HandleFunc("POST /v1/triggers/fire", s.handleFireTrigger)
	s.mux.HandleFunc("GET /v1/triggers/{id}", s.handleGetTrigger)
	s.mux.HandleFunc("PATCH /v1/triggers/{id}", s.handleUpdateTrigger)
	s.mux.HandleFunc("DELETE /v1/triggers/{id}", s.handleDeleteTrigger)
	s.mux.HandleFunc("POST /v1/triggers/{id}/enable", s.handleEnableTrigger)
	s.mux.HandleFunc("POST /v1/triggers/{id}/disable", s.handleDisableTrigger)
	s.mux.HandleFunc("POST /v1/triggers/{id}/rotate", s.handleRotateTrigger)
	s.mux.HandleFunc("POST /v1/triggers/{id}/reveal", s.handleRevealTrigger)
	s.mux.HandleFunc("GET /v1/filesystem-watchers", s.handleListFilesystemWatchers)
	s.mux.HandleFunc("POST /v1/filesystem-watchers", s.handleCreateFilesystemWatcher)
	s.mux.HandleFunc("GET /v1/filesystem-watchers/{id}", s.handleGetFilesystemWatcher)
	s.mux.HandleFunc("PATCH /v1/filesystem-watchers/{id}", s.handleUpdateFilesystemWatcher)
	s.mux.HandleFunc("DELETE /v1/filesystem-watchers/{id}", s.handleDeleteFilesystemWatcher)
	s.mux.HandleFunc("POST /v1/filesystem-watchers/{id}/enable", s.handleEnableFilesystemWatcher)
	s.mux.HandleFunc("POST /v1/filesystem-watchers/{id}/disable", s.handleDisableFilesystemWatcher)
	s.mux.HandleFunc("GET /v1/trigger-sets", s.handleListTriggerSets)
	s.mux.HandleFunc("POST /v1/trigger-sets", s.handleCreateTriggerSet)
	s.mux.HandleFunc("GET /v1/trigger-sets/{id}", s.handleGetTriggerSet)
	s.mux.HandleFunc("PATCH /v1/trigger-sets/{id}", s.handleRetargetTriggerSet)
	s.mux.HandleFunc("DELETE /v1/trigger-sets/{id}", s.handleDeleteTriggerSet)
	s.mux.HandleFunc("POST /v1/trigger-sets/{id}/enable", s.handleEnableTriggerSet)
	s.mux.HandleFunc("POST /v1/trigger-sets/{id}/disable", s.handleDisableTriggerSet)
	s.mux.HandleFunc("POST /v1/trigger-sets/{id}/rotate", s.handleRotateTriggerSet)
	s.mux.HandleFunc("POST /v1/trigger-sets/{id}/reveal", s.handleRevealTriggerSet)

	s.mux.HandleFunc("GET /v1/chains", s.handleListChains)
	s.mux.HandleFunc("POST /v1/chains", s.handleCreateChain)
	s.mux.HandleFunc("GET /v1/chains/{id}", s.handleGetChain)
	s.mux.HandleFunc("PATCH /v1/chains/{id}", s.handleUpdateChain)
	s.mux.HandleFunc("DELETE /v1/chains/{id}", s.handleDeleteChain)

	s.mux.HandleFunc("GET /v1/groups", s.handleListGroups)
	s.mux.HandleFunc("POST /v1/groups", s.handleCreateGroup)
	s.mux.HandleFunc("GET /v1/groups/{id}", s.handleGetGroup)
	s.mux.HandleFunc("PATCH /v1/groups/{id}", s.handleUpdateGroup)
	s.mux.HandleFunc("DELETE /v1/groups/{id}", s.handleDeleteGroup)
	s.mux.HandleFunc("POST /v1/groups/{id}/enable", s.handleEnableGroup)
	s.mux.HandleFunc("POST /v1/groups/{id}/disable", s.handleDisableGroup)

	s.mux.HandleFunc("POST /v1/schedules/preview", s.handlePreview)

	s.mux.HandleFunc("GET /v1/runs", s.handleListRuns)
	s.mux.HandleFunc("GET /v1/runs/active", s.handleListActiveRuns)
	s.mux.HandleFunc("GET /v1/runs/{id}", s.handleGetRun)

	s.mux.HandleFunc("GET /v1/notification-channels", s.handleListNotificationChannels)
	s.mux.HandleFunc("POST /v1/notification-channels", s.handleCreateNotificationChannel)
	s.mux.HandleFunc("GET /v1/notification-channels/{id}", s.handleGetNotificationChannel)
	s.mux.HandleFunc("PATCH /v1/notification-channels/{id}", s.handleUpdateNotificationChannel)
	s.mux.HandleFunc("DELETE /v1/notification-channels/{id}", s.handleDeleteNotificationChannel)
	s.mux.HandleFunc("POST /v1/notification-channels/{id}/enable", s.handleEnableNotificationChannel)
	s.mux.HandleFunc("POST /v1/notification-channels/{id}/disable", s.handleDisableNotificationChannel)
	s.mux.HandleFunc("POST /v1/notification-channels/{id}/rotate", s.handleRotateNotificationChannel)
	s.mux.HandleFunc("POST /v1/notification-channels/{id}/test", s.handleTestNotificationChannel)
	s.mux.HandleFunc("GET /v1/notification-deliveries", s.handleListNotificationDeliveries)
	s.mux.HandleFunc("GET /v1/tasks/{id}/notifications", s.handleTaskNotifications)
	s.mux.HandleFunc("PUT /v1/tasks/{id}/notifications", s.handleTaskNotifications)
	s.mux.HandleFunc("GET /v1/tasks/{id}/notifications/effective", s.handleEffectiveTaskNotifications)
	s.mux.HandleFunc("GET /v1/groups/{id}/notifications", s.handleGroupNotifications)
	s.mux.HandleFunc("PUT /v1/groups/{id}/notifications", s.handleGroupNotifications)
	s.mux.HandleFunc("GET /v1/alerts", s.handleListAlerts)
	s.mux.HandleFunc("POST /v1/alerts/{id}/ack", s.handleAckAlert)
	s.mux.HandleFunc("GET /v1/logs", s.handleListLogs)

	s.mux.HandleFunc("GET /v1/calendar", s.handleCalendar)
	s.mux.HandleFunc("GET /v1/events", s.handleEvents)
	s.mux.HandleFunc("GET /v1/access/actors", s.handleListActors)
	s.mux.HandleFunc("POST /v1/access/actors", s.handleCreateActor)
	s.mux.HandleFunc("PATCH /v1/access/actors/{id}", s.handleUpdateActor)
	s.mux.HandleFunc("POST /v1/access/actors/{id}/revoke", s.handleRevokeActor)
	s.mux.HandleFunc("GET /v1/access/pairings", s.handleListPairings)
	s.mux.HandleFunc("POST /v1/access/pairings", s.handleCreatePairing)
	s.mux.HandleFunc("POST /v1/access/pairings/{id}/cancel", s.handleCancelPairing)
	s.mux.HandleFunc("GET /v1/access/credentials", s.handleListCredentials)
	s.mux.HandleFunc("POST /v1/access/credentials/{id}/rotate", s.handleRotateCredential)
	s.mux.HandleFunc("POST /v1/access/credentials/{id}/revoke", s.handleRevokeCredential)
	s.mux.HandleFunc("GET /v1/audit", s.handleListAudit)
	s.mux.HandleFunc("GET /v1/audit/export", s.handleExportAudit)

	// Fallback: unmatched routes return the consistent error envelope.
	s.mux.HandleFunc("/", s.handleNotFound)
}

func (s *Server) handleNotFound(w http.ResponseWriter, r *http.Request) {
	writeError(w, http.StatusNotFound, CodeNotFound, "", "no such endpoint: "+r.Method+" "+r.URL.Path)
}

// HealthResponse is returned by GET /v1/health.
type HealthResponse struct {
	Status  string `json:"status"`
	Version string `json:"version"`
}

func (s *Server) handleHealth(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, HealthResponse{Status: "ok", Version: buildinfo.Version})
}

func (s *Server) handleRuntimeInfo(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, s.runtime)
}

// ---- response helpers ---------------------------------------------------

// APIError is the consistent error envelope: {"error": {...}}.
type APIError struct {
	Error ErrorBody `json:"error"`
}

// ErrorBody describes a single error.
type ErrorBody struct {
	Code    string `json:"code"`
	Field   string `json:"field,omitempty"`
	Message string `json:"message"`
}

// Error codes used across the API.
const (
	CodeValidation       = "validation_failed"
	CodeNotFound         = "not_found"
	CodeConflict         = "conflict"
	CodeInternal         = "internal"
	CodeForbidden        = "forbidden"
	CodeAuditUnavailable = "audit_unavailable"
)

func writeJSON(w http.ResponseWriter, status int, v any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(v)
}

// writeError emits the error envelope with the given HTTP status.
func writeError(w http.ResponseWriter, status int, code, field, msg string) {
	writeJSON(w, status, APIError{Error: ErrorBody{Code: code, Field: field, Message: msg}})
}